+++ TextEditor/src/Command/MoveCursorCommand/MoveCursorCommand.cpp
@@ -2,7 +2,7 @@
 
 MoveCursorCommand::~MoveCursorCommand() = default;
 
-MoveCursorCommand::MoveCursorCommand(TextEditor& textEditor, const int start, const int end)
+MoveCursorCommand::MoveCursorCommand(TextEditor& textEditor, const unsigned long long start, const unsigned long long end)
     : textEditor(textEditor),
       start(start),
       end(end)
